backend/requirements.txt
flask
flask-sqlalchemy
oracledb
python-docx
python-dotenv